Move topology displayTabs to own function
