+++ src/views/TopologyDisplayView.vue
@@ -118,7 +118,6 @@ import type { WebOcTopologyDisplayConfig } from '@deltares/fews-pi-requests'
 
 import { computed, ref, StyleValue, watch, watchEffect } from 'vue'
 import {
-  type RouteLocationNamedRaw,
   onBeforeRouteUpdate,
   RouteLocationNormalized,
   useRoute,
@@ -131,6 +130,7 @@ import { useDisplay } from 'vuetify'
 import { useNodesStore } from '@/stores/nodes'
 import { nodeButtonItems, recursiveUpdateNode } from '@/lib/topology/nodes'
 import { useDownloadDialogStore } from '@/stores/downloadDialog'
+import { displayTabsForNode, type DisplayTab } from '@/lib/topology/displayTabs.js' 
 
 interface Props {
   nodeId?: string | string[]
@@ -139,16 +139,6 @@ interface Props {
   locationId?: string
   latitude?: string
   longitude?: string
-}
-
-interface DisplayTab {
-  type: 'charts' | 'map' | 'reports' | 'schematic-status-display'
-  id: string
-  title: string
-  href?: string
-  target?: string
-  to: RouteLocationNamedRaw
-  icon: string
 }
 
 const props = defineProps<Props>()
@@ -274,99 +264,6 @@ function updateItems(): void {
   }
 }
 
-function displayTabsForNode(leafNode: TopologyNode, parentNodeId?: string) {
-  const activeNodeId = leafNode.id
-  const timeseriesTabId = `${activeNodeId}-timeseries`
-  const reportsTabId = `${activeNodeId}-reports`
-  const spatialTabId = `${activeNodeId}-spatial`
-  const ssdTabId = `${activeNodeId}-ssd`
-  const _displayTabs: DisplayTab[] = []
-  if (
-    leafNode.gridDisplaySelection !== undefined ||
-    leafNode.filterIds !== undefined
-  ) {
-    _displayTabs.push({
-      type: 'map',
-      id: spatialTabId,
-      title: 'Map',
-      to: {
-        name: 'TopologySpatialDisplay',
-        params: {
-          nodeId: parentNodeId ? [parentNodeId, leafNode.id] : leafNode.id,
-          layerName: leafNode.gridDisplaySelection?.plotId,
-        },
-      },
-      icon: 'mdi-map',
-    })
-  }
-  if (
-    leafNode.displayGroups !== undefined ||
-    leafNode.displayId !== undefined ||
-    (leafNode.plotId != undefined && leafNode.locationIds != undefined)
-  ) {
-    _displayTabs.push({
-      type: 'charts',
-      id: timeseriesTabId,
-      title: 'Charts',
-      to: {
-        name: 'TopologyTimeSeries',
-        params: {
-          nodeId: parentNodeId ? [parentNodeId, leafNode.id] : leafNode.id,
-        },
-      },
-      icon: 'mdi-chart-multiple',
-    })
-  }
-  if (
-    leafNode.filterIds !== undefined &&
-    leafNode.filterIds.length == 1 &&
-    leafNode.dataDownloadDisplay !== undefined
-  ) {
-    _displayTabs.push({
-      type: 'charts',
-      id: timeseriesTabId,
-      title: 'Download',
-      to: {
-        name: 'TopologyDataDownload',
-        params: {
-          nodeId: parentNodeId ? [parentNodeId, leafNode.id] : leafNode.id,
-        },
-      },
-      icon: 'mdi-download',
-    })
-  }
-  if (leafNode.reportDisplay?.reports.length) {
-    _displayTabs.push({
-      type: 'reports',
-      id: reportsTabId,
-      title: 'Reports',
-      to: {
-        name: 'TopologyReports',
-        params: {
-          nodeId: parentNodeId ? [parentNodeId, leafNode.id] : leafNode.id,
-        },
-      },
-      icon: 'mdi-file-document',
-    })
-  }
-  if (leafNode.scadaPanelId !== undefined) {
-    _displayTabs.push({
-      type: 'schematic-status-display',
-      id: ssdTabId,
-      title: 'Schematic',
-      to: {
-        name: 'TopologySchematicStatusDisplay',
-        params: {
-          nodeId: parentNodeId ? [parentNodeId, leafNode.id] : leafNode.id,
-          panelId: leafNode.scadaPanelId,
-        },
-      },
-      icon: 'mdi-view-dashboard',
-    })
-  }
-  return _displayTabs
-}
-
 watch(nodes, updateItems)
 watch(thresholds, updateItems)
 
@@ -408,9 +305,7 @@ watchEffect(() => {
 
   // Create the displayTabs for the active node.
   if (node === undefined) return
-  const _displayTabs = displayTabsForNode(node, parentNodeIdNodeId)
-  displayTabs.value = _displayTabs
-
+  displayTabs.value = displayTabsForNode(node, parentNodeIdNodeId)
   externalLink.value = node.url
 })
